Fix time truncation in AffineSchedule with integer-typed offset
Drop explicit dtype cast on t in coeffs_at to prevent silent truncation
when offset is integer-typed (e.g. AffineSchedule(offset=1, slope=0.5)).

+++ src/vmc/operators/time_dependent.py
@@ -70,7 +70,7 @@ def coeffs_at(schedule: TermCoefficientSchedule, t: float | jax.Array) -> jax.Ar
 
 @coeffs_at.dispatch
 def coeffs_at(schedule: AffineSchedule, t: float | jax.Array) -> jax.Array:
-    return schedule.offset + jnp.asarray(t, dtype=schedule.offset.dtype) * schedule.slope
+    return schedule.offset + jnp.asarray(t) * schedule.slope
 
 
 @dispatch
